requirements.txt
# Required dependencies
flask==3.1.1
flask-cors==6.0.1
opencv-python==4.12.0.88
numpy==2.2.6
Pillow>=9.1.0
pytesseract>=0.3.10

# Optional dependencies for enhanced OCR
# easyocr>=1.7.0  # Uncomment to enable EasyOCR fallback

# For math solving
sympy>=1.12

# For AI explanations (optional)
# google-generativeai>=0.3.0  # Uncomment to enable Gemini API 